v6

+++ app/src/main/java/com/example/recycler/fragment/FragmentDownload.java
@@ -93,12 +93,15 @@ public class FragmentDownload extends Fragment implements HistoryAdapter.ClickLi
     public void onDestroy() {
         super.onDestroy();
         for (Article deletedItem:listDelete){
-            List<Description> listDescriptions = appDataBase.descriptionDao().getListImageDescription(deletedItem.getId());
-            for (Description description : listDescriptions) {
-                deleteFIle(description.getContent());
+            if(appDataBase.articleDao().getListName(deletedItem.getLinkImage()).size()<2){
+                List<Description> listDescriptions = appDataBase.descriptionDao().getListImageDescription(deletedItem.getId());
+                for (Description description : listDescriptions) {
+                    deleteFIle(description.getContent());
 
+                }
+                deleteFIle(deletedItem.getLinkImage());
             }
-            deleteFIle(deletedItem.getLinkImage());
+
             appDataBase.descriptionDao().deleteAll(deletedItem.getId());
             appDataBase.articleDao().delete(deletedItem);
         }
